% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfile.R
\name{celsius_to_fahrenheit}
\alias{celsius_to_fahrenheit}
\title{Μετατροπή Celsius σε Fahrenheit}
\usage{
celsius_to_fahrenheit(celsius)
}
\arguments{
\item{celsius}{Αριθμητική τιμή σε βαθμούς Κελσίου.}
}
\value{
Η θερμοκρασία σε βαθμούς Φαρενάιτ.
}
\description{
Δέχεται μια τιμή σε βαθμούς Κελσίου και την μετατρέπει σε Φαρενάιτ.
}
\examples{
celsius_to_fahrenheit(0)
celsius_to_fahrenheit(25)
}
