% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cols_in.R
\name{clean_cols_in}
\alias{clean_cols_in}
\title{Make column names easier to work with}
\usage{
clean_cols_in(column_names, case = c("upper", "lower", "proper"))
}
\value{
A vector of characters.
}
\description{
Substitutes symbols with spaces, removes double spaces and trailing
white space, replaces spaces with underscores and slashes with periods;
optionally changes string case.
}
\examples{
column_names <- c("Amount$", "Payment Date\r\n(mm/dd/yyyy)")
clean_cols_in(column_names, "lower")
}
\seealso{
\code{\link[=clean_cols_out]{clean_cols_out()}}
}
