% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cols_out.R
\name{clean_cols_out}
\alias{clean_cols_out}
\title{Improve the look of column names prior to export}
\usage{
clean_cols_out(column_names, case = c("upper", "lower", "title"))
}
\arguments{
\item{column_names}{A character vector of column names.}

\item{case}{An optional argument to change the case of the
strings provided to \code{"upper", "lower",} or \code{"title"}.}
}
\value{
A vector of characters.
}
\description{
Replaces underscores with spaces and periods with slashes;
optionally changes string case.
}
\examples{
column_names <- c("amount", "payment_date_mm.dd.yyyy")
column_names
clean_cols_out(column_names, "upper")
}
\seealso{
\code{\link[=clean_cols_in]{clean_cols_in()}}
}
