\name{ContaminantsSeq}
\alias{ContaminantsSeq}
\docType{data}
\title{Common contaminants in mass spectrometry}
\description{Protein sequences of common contaminants in mass spectrometry.}
\usage{data(ContaminantsSeq)}
\format{
\verb{  A AAStringSet instance of length 248
      width seq                               names               
  [1]   231 FPTDDDDKIVGGYTC...VCNYVNWIQQTIAAN CONT_P00761 SWISS...
  [2]   540 MSRQFTYKSGAAAKG...GKTLALSSPTKKTMR CONT_Q32MB2 TREMB...
  [3]   594 MTSVGVFSDMLNGCG...SSSKIISTTTLNKRR CONT_P19013 SWISS...
  [4]   521 MSLSPCRAQRGFSAR...LKKTVESSLKTSITY CONT_Q7RTT2 TREMB...
  [5]   653 MKRICGSLLLLGLSI...DLSSEAAQRAPGSCG CONT_P15636 SWISS...
  ...   ... ...
[244]   451 MRECISIHVGQAGVQ...VDSVEGEGEEEGEEY CONT_ENSEMBL:ENSB...
[245]   452 MWAVLSLPLACLLAQ...RTVLFLGRVVDPTLL CONT_ENSEMBL:ENSB...
[246]   591 MSFGNWKPTVVVQGI...KRRPPPTILLSTDLQ CONT_ENSEMBL:ENSB...
[247]  1477 MGKNKLLYPSLTLLL...VAEYSAPCSKDIGNA CONT_ENSEMBL:ENSB...
[248]   604 MKQLKLTGFVIFFFF...FKEESPKLAAKSQAA CONT_REFSEQ:XP_58...
}}
\examples{
data(ContaminantsSeq, package="RBDmap")
ContaminantsSeq
}
\keyword{datasets}
