% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_na_rows.R
\name{drop_na_rows}
\alias{drop_na_rows}
\title{Drop NA rows}
\usage{
drop_na_rows(df)
}
\arguments{
\item{df}{A data frame or tibble.}
}
\value{
A data frame or tibble.
}
\description{
Returns a data frame without rows containing only NA values.
}
\examples{
sample_data <- rbind(head(iris, 5), rep(NA, 5))
sample_data

drop_na_rows(sample_data)
}
\seealso{
\code{\link[=drop_na_cols]{drop_na_cols()}}
}
