% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_matching.R
\name{files_matching}
\alias{files_matching}
\title{Locate files in a directory}
\usage{
files_matching(path, file_type, name_contains = NULL, ignore.case = TRUE)
}
\arguments{
\item{path}{The path to the directory to locate files.}

\item{file_type}{A character vector indicating a file extension.}

\item{name_contains}{An optional character vector to filter for matching file names.}

\item{ignore.case}{If \code{TRUE}, \code{name_contains} case is ignored when filtering matching file names.}
}
\value{
A three column tibble.
}
\description{
Search a directory for a given file type and name, return
a tibble of matching file information sorted by modified date.
}
\examples{
search_directory <- getwd() # change prior to running example

# search for pdf files
files_matching(search_directory, ".pdf")

# search for Excel files with "report" included in the title
files_matching(search_directory, ".xlsx", name_contains = "report")

df_matches <- files_matching(search_directory, ".R")
df_matches

latest_match <- df_matches$path[1]
latest_match
}
