% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_page_list.R
\name{flatten_page_lines}
\alias{flatten_page_lines}
\title{Structure text using return characters}
\usage{
flatten_page_lines(page_list, rm_empty_str = TRUE)
}
\value{
A one column tibble.
}
\description{
Converts a list of characters to a tibble where each row represents
a line of text. Primarily for use with \code{\link[pdftools:pdftools]{pdftools::pdf_text()}} or
\code{\link[pdftools:pdf_ocr]{pdftools::pdf_ocr_text()}} when return character "\\n" separates data
in a meaningful way.
}
\examples{
\code{ex_page1 <- paste0(stringr::fruit[1:5], collapse = "\n")}
\code{ex_page2 <- paste0(stringr::fruit[6:10], collapse = "\n")}
ex_page_list <- list(ex_page1, ex_page2)
ex_page_list

flatten_page_lines(ex_page_list)
}
