\name{getBindingCategory}
\alias{getBindingCategory}
\title{Get binding category of peptides}
\description{Get the binding category of peptides}
\usage{
getBindingCategory(E, efit, maxFDR = 0.01)
}
\arguments{
  \item{E}{}
  \item{efit}{}
  \item{maxFDR}{}
}
\details{
}
\value{
}
\author{Bernd Fischer}
\examples{
}
%\keyword{ ~kwd1 }
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
