% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casnet_SOURCE_jamovi.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Whole number?}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{A number}

\item{tol}{The tolerance (\code{default = Machine$double.eps^0.5})}
}
\value{
TRUE or FALSE
}
\description{
Whole number?
}
