\name{RBDmap-package}
\alias{RBDmap-package}
\alias{RBDmap}
\docType{package}
\title{RBDmap}
\description{RBDmap functions and datasets.}
\details{
\tabular{ll}{
Package: \tab RBDmap\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
License: \tab Artistic-2.0\cr
LazyLoad: \tab true\cr
Imports: \tab Biostrings, Biobase, hash, biomaRt\cr
Depends: \tab R (>= 2.10)\cr
Suggests: \tab grid, BiocStyle, RColorBrewer\cr
SystemRequirements: \tab GNU make\cr
biocViews: \tab Infrastructure\cr
Built: \tab R 3.0.2; ; 2013-12-06 07:35:28 UTC; unix\cr
}

Index:
\preformatted{
ContaminantsSeq         Common contaminants in mass spectrometry
downloadProteinAnnotation
                        Download protein annotation
getComplexity           Local sequence complexity
}
}
\author{
Bernd Fischer

Maintainer: Bernd Fischer <bernd.fischer@embl.de>
}
\references{
}
\keyword{ package }
