% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_url_zip.R
\name{read_url_zip}
\alias{read_url_zip}
\title{Load data from online zipped folders}
\usage{
read_url_zip(
  url,
  file_type = "csv",
  file_name = "",
  file_index = 1,
  reader = NULL
)
}
\arguments{
\item{url}{The download link to a zipped folder online.}

\item{file_type}{A character vector indicating a file extension.}

\item{file_name}{An optional string to specify which file to read by name.}

\item{file_index}{An optional index number to specify which of the files to read when \code{file_name} isn't specified.}

\item{reader}{A function specified by a string that is called to load the file into R.}
}
\value{
A tibble when the file type is a csv, tsv, xlsx, or xls, and a character vector when the file type is pdf or txt.
}
\description{
Downloads the zip folder from the provided url, unzips it to a temporary
location, reads the first csv into R as a tibble, then deletes the downloaded
zip folder and the temporary files.
}
\details{
The file to read from within the folder can be altered with \code{file_type},
\code{file_name}, and \code{file_index}. If the file to read is not a "csv", "tsv",
"xlsx", "xls", "pdf", or "txt" file type, a function to read the file with
must be specified by \code{reader}.
}
\examples{
\dontrun{
url <- "url.zip"

# load the first csv
data <- read_url_zip(url)

# load the xlsx file "data 2020.xlsx"
data <- read_url_zip(url, file_type = "xlsx", file_name = "data 2020")

# load the 2nd txt file and use the `read_csv` function instead of the default txt reader `readLines`
data <- read_url_zip(url, file_type = "txt", file_index = 2, reader = "read_csv")

}
}
