% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperature_conversion.R
\name{temperature_conversion}
\alias{temperature_conversion}
\title{Convert temperature representations among common units}
\usage{
temperature_conversion(input_unit, output_unit, input_temp)
}
\arguments{
\item{input_unit}{Unit of the input temperature. The unit should be one of the strings from "C', "K", "F", "Celsius", "Kelvin" and "Fahrenheit".}

\item{output_unit}{Unit of the output temperature. The unit should also be one of the string from "C', "K", "F", "Celsius", "Kelvin" and "Fahrenheit". Same input and output units are not allowed, and will generate error messages.}

\item{input_temp}{Input temperature before conversion. The input temperature can come from manual input to the function argument, or it can come from calling on data columns from datasets.}
}
\value{
Temperature after conversion that is in the specified output unit.
}
\description{
The function temperature_conversion() converts a numerical representation of temperature among three common units: Fahrenheit, Celsius and Kelvin.
}
\examples{
temperature_conversion("Celsius", "Kelvin", 50)
temperature_conversion("K", "F", 50)
temperature_conversion("Celsius", "F", 50)
}
