% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temps}
\alias{temps}
\alias{recpre_long}
\title{Data}
\format{
A data frame with 60 rows and 10 variables:
\describe{
  \item{id}{Subject identifier}
  \item{trial_num}{order in which the experimental trial was completed}
  \item{trial_condition}{Environmental condition and metabolic heat production}
  \item{tod}{Time of Day}
  \item{trec_pre}{Rectal temperature before the beginning of the trial}
  \item{trec_post}{Rectal temperature at the end of the trial}
  \item{trec_delta}{Change in rectal temperature}
  \item{teso_pre}{Esophageal temperature before the beginning of the trial}
  \item{teso_post}{Esophageal temperature at the end of the trial}
  \item{teso_delta}{Change in esophageal temperature}

}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30 rows and 6 columns.
}
\source{
Ravanelli N, Jay O. The Change in Core Temperature and Sweating Response during Exercise Are Unaffected by Time of Day within the Wake Period. Med Sci Sports Exerc. 2020 Dec 1. doi: 10.1249/MSS.0000000000002575. Epub ahead of print. PMID: 33273272.
}
\usage{
temps

recpre_long
}
\description{
A dataset from a study on the reliability of human body temperature at different times of day before and after exercise.
}
\keyword{datasets}
