% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_cols_out.R
\name{clean_cols_out}
\alias{clean_cols_out}
\title{Improve the look of column names prior to export}
\usage{
clean_cols_out(column_names, case = c("upper", "lower", "proper"))
}
\value{
A vector of characters.
}
\description{
Replaces underscores with spaces and periods with slashes;
optionally changes string case.
}
\examples{
column_names <- c("AMOUNT", "PAYMENT_DATE_MM.DD.YYYY")
clean_cols_out(column_names, "proper")
}
\seealso{
{
\code{\link[=clean_cols_in]{clean_cols_in()}}
}
}
