% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files_matching.R
\name{files_matching}
\alias{files_matching}
\title{Search a directory for files}
\usage{
files_matching(path, file_type, name_contains = NULL, ignore.case = T)
}
\value{
An integer vectors or list of integer vectors.
}
\description{
todo
}
\examples{
files_matching(getwd(), "pdf")
files_matching(getwd(), "xlsm", name_contains = "a")
}
