% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vary-package.R
\docType{package}
\name{vary-package}
\alias{vary}
\alias{vary-package}
\title{vary: Manage Unstructured Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Methods to reliably structure data when the source is expected to vary in format or have no structure at all.
}
\author{
\strong{Maintainer}: Chris Ulch \email{culch@ualberta.ca}

}
\keyword{internal}
