% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vary-package.R
\docType{package}
\name{vary-package}
\alias{vary}
\alias{vary-package}
\title{vary: Manage Unstructured Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Methods to reliably structure data when the source format or load point is expected to vary over time; or when inputs lack structure completely and are not immediately compatible with typical data manipulation packages. For situations where connection points are pdfs or user-edited Excel sheets, and table dimensions and naming conventions need to be dynamically set using pattern recognition.
}
\author{
\strong{Maintainer}: Chris Ulch \email{culch@ualberta.ca}

}
\keyword{internal}
