% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vary-package.R
\docType{package}
\name{vary-package}
\alias{vary}
\alias{vary-package}
\title{vary: Manage Unstructured Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Methods to reliably structure data when the source format or load point is expected to vary over time, or when inputs lack structure completely and are not immediately compatible with typical data manipulation packages (ex. OCR output, user-edited Excel sheets). Streamlines a few string methods to make field verification with pattern recognition easier.
}
\seealso{
Structure Data

\itemize{
\item \code{\link[=which_rows]{which_rows()}}
}

\itemize{
\item \code{\link[=fuzzy_match]{fuzzy_match()}}
}

\itemize{
\item \code{\link[=fuzzy_rename]{fuzzy_rename()}}
}

\itemize{
\item \code{\link[=flatten_page_lines]{flatten_page_lines()}}
}

\itemize{
\item \code{\link[=drop_na_rows]{drop_na_rows()}}
}

\itemize{
\item \code{\link[=drop_na_cols]{drop_na_cols()}}
}

Locate Files

\itemize{
\item \code{\link[=get_downloads_folder]{get_downloads_folder()}}
}

\itemize{
\item \code{\link[=get_desktop_folder]{get_desktop_folder()}}
}

\itemize{
\item \code{\link[=files_matching]{files_matching()}}
}

Other \emph{(possible future relocation)}

\itemize{
\item \code{\link[=clean_cols_in]{clean_cols_in()}}
}

\itemize{
\item \code{\link[=clean_cols_out]{clean_cols_out()}}
}

\itemize{
\item \code{\link{\%notin\%}}
}
}
\author{
\strong{Maintainer}: Chris Ulch \email{culch@ualberta.ca}

}
\keyword{internal}
