% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vary-package.R
\docType{package}
\name{vary-package}
\alias{vary}
\alias{vary-package}
\title{vary: Manage Unstructured Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Methods to automate the loading of semi-structured data (ex. user modified files, OCR output) which are reliable enough to form a process around, but vary too much to immediately work with using typical data manipulation packages.
}
\seealso{
Structure Data

\itemize{
\item \code{\link[=which_rows]{which_rows()}}
}

\itemize{
\item \code{\link[=fuzzy_match]{fuzzy_match()}}
}

\itemize{
\item \code{\link[=fuzzy_rename]{fuzzy_rename()}}
}

\itemize{
\item \code{\link[=flatten_pages]{flatten_pages()}}
}

\itemize{
\item \code{\link[=page_search]{page_search()}}
}

\itemize{
\item \code{\link[=drop_na_rows]{drop_na_rows()}}
}

\itemize{
\item \code{\link[=drop_na_cols]{drop_na_cols()}}
}

Locate Files

\itemize{
\item \code{\link[=get_downloads_folder]{get_downloads_folder()}}
}

\itemize{
\item \code{\link[=get_desktop_folder]{get_desktop_folder()}}
}

\itemize{
\item \code{\link[=files_matching]{files_matching()}}
}

Other \emph{(possible future relocation)}

\itemize{
\item \code{\link{\%notin\%}}
}
}
\author{
\strong{Maintainer}: Chris Ulch \email{culch@ualberta.ca}

}
\keyword{internal}
