% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_rows.R
\name{which_rows}
\alias{which_rows}
\title{Return matching row indices}
\usage{
which_rows(
  df,
  contain_strings = NULL,
  scan_cols = 1:ncol(df),
  all_strings = TRUE,
  case_sensitive = TRUE,
  lack_na = TRUE,
  flatten = TRUE
)
}
\value{
An integer vector or list of integer vectors
}
\description{
Replaces underscores with spaces and periods with slashes;
optionally changes string case.
}
\examples{
column_names <- c("amount", "payment_date_mm.dd.yyyy")
which_rows(column_names, "upper")
}
